requirements.txt
# Core Framework
crewai[tools,google-genai]>=0.11.0

# Web Framework
fastapi>=0.110.0
uvicorn[standard]>=0.29.0

# PDF Processing
pypdf>=4.0.0

# Environment Variables
python-dotenv>=1.0.0

# Database (Bonus Feature)
sqlalchemy>=2.0.0

# Celery Queue Worker (Bonus Feature)
celery[redis]>=5.3.0
redis>=5.0.0